GA3/q02_llm_structured_output_fastapi_sentiment/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.55.0
